+++ src/backend/core/api/utils.py
@@ -11,6 +11,29 @@ import botocore
 from rest_framework.throttling import BaseThrottle
 
 
+def filter_root_paths(paths, skip_sorting=False):
+    """
+    Filters root paths from a list of paths representing a tree structure.
+    A root path is defined as a path that is not a prefix of any other path.
+
+    Args:
+        paths (list of str): The list of paths.
+
+    Returns:
+        list of str: The filtered list of root paths.
+    """
+    if not skip_sorting:
+        paths.sort()
+
+    root_paths = []
+    for path in paths:
+        # If the current path is not a prefix of the last added root path, add it
+        if not root_paths or not path.startswith(root_paths[-1]):
+            root_paths.append(path)
+
+    return root_paths
+
+
 def generate_s3_authorization_headers(key):
     """
     Generate authorization headers for an s3 object.